==> 01_merging.ipynb <==
# Merging Data

## Import all necessary libraries

import pandas as pd
import numpy as np

### Step 1: `Characteristics` (2005-2023)

char_21 = pd.read_csv(
    './data/raw/caracteristiques 2005-2021.csv',
    encoding='ISO-8859-1',
    index_col=0,
    low_memory=False,
)

char_22 = pd.read_csv('./data/raw/carcteristiques-2022.csv', sep=';')

char_23 = pd.read_csv('./data/raw/caract-2023.csv', sep=';')

print('#'*10, '2005-2021')
print(char_21.head(), end='\n\n')
print('#'*10, '2022')
print(char_22.head(), end='\n\n')
print('#'*10, '2023')
print(char_23.head(), end='\n\n')

There are different column names across data sets that address the same problem. For `long` and `lat`, I find the wrong decimal sign (`,` instead of `.`)

char_21.rename(columns={'annee': 'an'}, inplace=True)
char_22.rename(columns={'Accident_Id': 'num_acc'}, inplace=True)
char_23.rename(columns={'Num_Acc': 'num_acc'}, inplace=True)

char = pd.concat(
    [char_21, char_22, char_23],
    ignore_index=True
)

char.head()

It is clear that the `lat` and `long` features have a wrong decimal sign and that (as seen above) the decimal sign is missing. However, I deal with this in the feature engineering step.

### Step 2: Places `lieux` 2005-2023

places_19 = pd.read_csv(
    './data/raw/lieux 2005-2019.csv', 
    low_memory=False, 
    index_col=0
)

places_20 = pd.read_csv(
    './data/raw/lieux-2020.csv', 
    sep=';', 
    low_memory=False
)

places_21 = pd.read_csv(
    './data/raw/lieux-2021.csv', 
    sep=';', 
    low_memory=False
)

places_22 = pd.read_csv(
    './data/raw/lieux-2022.csv', 
    sep=';', 
    low_memory=False
)

places_23 = pd.read_csv(
    './data/raw/lieux-2023.csv', 
    sep=';', 
    low_memory=False
)

print('#'*10, '2005-2019')
print(places_19.head(), end='\n\n')
print('#'*10, '2020')
print(places_20.head(), end='\n\n')
print('#'*10, '2021')
print(places_21.head(), end='\n\n')
print('#'*10, '2022')
print(places_22.head(), end='\n\n')
print('#'*10, '2023')
print(places_23.head())

Again, I have to rename two columns in the file from 2005 to 2021 (`places_21`) and I have to add the year for the files in 2022 (`places_22`) and 2023 (`places_23`). Furthermore, I have a variable `env1` in the file from 2005 to 2021, which is not reported in the other data sets, therefore, I drop it here. 

places_19 = places_19.drop(columns=['env1'], axis=1)
places_19.rename(columns={'annee': 'an'}, inplace=True)

places_20.rename(columns={'Num_Acc': 'num_acc'}, inplace=True)
places_20['an'] = 2020

places_21.rename(columns={'Num_Acc': 'num_acc'}, inplace=True)
places_21['an'] = 2021

places_22.rename(columns={'Num_Acc': 'num_acc'}, inplace=True)
places_22['an'] = 2022

places_23.rename(columns={'Num_Acc': 'num_acc'}, inplace=True)
places_23['an'] = 2023

places = pd.concat(
    [places_19, places_20, places_21, places_22, places_23], 
    ignore_index=True
)

places.head()

### Step 3: `usagers` 2005-2023

From this data set that inherits information of each person involved, I only need the variable `grav` to calculate the overall accident's severity score. 

users_21 = pd.read_csv(
    './data/raw/usagers 2005-2021.csv',
    index_col=0, 
    low_memory=False
)

users_22 = pd.read_csv(
    './data/raw/usagers-2022.csv',
    sep=';', 
    low_memory=False
)
users_22['an'] = 2022

users_23 = pd.read_csv(
    './data/raw/usagers-2023.csv', 
    sep=';', 
    low_memory=False
)
users_23['an'] = 2023

Quick renaming the accident number to merge the files later on easier. 

print('#'*10, '2005-2009')
print(users_21.head(), end='\n\n')
print('#'*10, '2022')
print(users_22.head(), end='\n\n')
print('#'*10, '2023')
print(users_23.head(), end='\n\n')

users_21.rename(columns={'annee': 'an'}, inplace=True)
users_22.rename(columns={'Num_Acc': 'num_acc'}, inplace=True)
users_23.rename(columns={'Num_Acc': 'num_acc'}, inplace=True)

users_21 = pd.DataFrame(users_21[['grav', 'num_acc', 'an']])
users_22 = pd.DataFrame(users_22[['grav', 'num_acc', 'an']])
users_23 = pd.DataFrame(users_23[['grav', 'num_acc', 'an']])

print(users_21.head())
print(users_22.head())
print(users_23.head())

users = pd.concat(
    [users_21, users_22, users_23], 
    ignore_index=True
)

users.head()

# Step 4: Merging the three data frames

merged_df = pd.merge(
    char, places, 
    on=['num_acc', 'an'], 
    how='inner'
) 

merged_df = pd.merge(
    merged_df, users, 
    on=['num_acc', 'an'], 
    how='inner'
)

merged_df.head()

To make it a bit more intuitive, I change variable names to an English shortname. 

new_cols = {
    'mois': 'month',
    'jour': 'day',
    'hrmn': 'time',
    'lum': 'light_condit',
    'agg': 'location_type',
    'int': 'intersect_type',
    'atm': 'weather_cond',
    'col': 'collision_type',
    'com': 'municip_code',
    'adr': 'adr',
    'gps': 'gps_code',
    'lat': 'lat',
    'long': 'long',
    'dep': 'dep',
    'an': 'year',
    'catr': 'road_cat',
    'voie': 'road_num',
    'v1': 'road_num_index',
    'v2': 'v2',
    'circ': 'traffic_dir',
    'nbv': 'num_lanes',
    'pr': 'pr',
    'pr1': 'pr1',
    'vosp': 'res_lane',
    'prof': 'prof_road',
    'plan': 'plan_view',
    'lartpc': 'tpc_length',
    'larrout': 'width_road',
    'surf': 'surface',
    'infra': 'infra',
    'situ': 'acc_loc',
    'vma': 'max_speed',
    'place': 'pos_in_car',
    'catu': 'user_cat',
    'grav': 'gravity',
    'sexe': 'sex',
    'trajet': 'travel_aim',
    'locp': 'loc_pedest',
    'actp': 'act_pedest',
    'etatp': 'ped_num',
    'an_nais': 'birth_year',
    'num_veh': 'num_car',
    'id_vehicule': 'id_car',
    'secu1': 'secu1',
    'secu2': 'secu2',
    'secu3': 'secu3',
    'id_usager': 'id_user',
    'senc': 'dir_traffic',
    'catv': 'cat_car',
    'occutc': 'occutc',
    'obs': 'fixed_obj',
    'obsm': 'mobile_object',
    'choc': 'impact',
    'manv': 'last_move',
    'motor': 'engine_type',
}


merged_df.rename(new_cols, axis=1, inplace=True)

## Step 5: Selecting mainland France

In the feature `dep`, there is information in which department the accident happened. With this, I can exclude accidents that happened in oversea territories, since I want to focus on mainland France. Departments with values over $970$ are oversea departements. However, there are some departments 

merged_df['dep'].unique()

# drop corse
merged_df = merged_df[(merged_df['dep'] != '2A') & (merged_df['dep'] != '2B')]

# drop overseas 
merged_df.loc[:, 'dep'] = merged_df['dep'].astype(int)
merged_df = merged_df[merged_df['dep'] < 970]

There are still departments with three digits, they all end with zero. We cut off the zero since this might be an indication for subclassification in the department.

Finally, it's just saving the data. In the next notebook `02_cleaning.ipynb`, I will preprocess the data to make the features clean. Further processing is then done in notebook `04_feature-engineering.ipynb`.

merged_df.to_csv('./data/processed/merged.csv', index=False)
