==> 02_cleaning.ipynb <==
# Data Cleaning

In this step, I will take a look in each variable, to clean features before doing explorative analysis. 

## Import libraries + load data set

# libraries used
import pandas as pd
import numpy as np

# for geo manipulation
import geopandas as gpd
from shapely.geometry import Point

# loading data
data = pd.read_csv(
    'data/processed/merged.csv',
    low_memory=False
)

# Target creation

As I outlined in the `README.md`, the aim is to predict accident's overall severity score. Therefore, we need to built an index of severity based on individual gravity (`grav`). 

The formular to calculate the overall gravity is (see `README.md` for more information): 

$$Severity Score = \frac{1}{n} \sum_{i=1}^n is_i  + weight_{killed} * pk^{0.4} + weight_{hospital} * ph^{0.2}$$

with $is$: individual score
$pk$: persons killed
$ph$: persons in hospital

Hence, I need to create counts of `killed` and `hospital` per accident. 

For `gravity` values are: `-1` no information, `1` uninjured, `2` killed, `3` in hospital, `4` lightly injured. In the following, I first apply the values (in relation to AIS), sum the values over each accident and take the mean. Then I create two dummies `killed` and `hospital` that count the people killed and injured in hospital per accident. Finally, I apply the extra weights for the amount of person killed or in hospital per accident.

weights = {
    1: 0,  # Uninjured, but involved
    2: 6,  # killed
    3: 3.5,  # injured in hospitel
    4: 1,  # lightly injured
    -1: 0.0,  # no information, nothing added
}
# based on scale for accident severity (see report)

# applying weights to grav and save it in new feature
data['grav_weighted'] = data['gravity'].map(weights)

# summing up over accident
data['severity_score'] = data.groupby('num_acc')[
    'grav_weighted'].transform('mean')

# Dummy: killed
data['killed'] = [1 if gravity == 2 else 0 for gravity in data['gravity']]

# Sum of killed per accident
data['killed_n'] = data.groupby('num_acc')['killed'].transform('sum')

# Dummy: Hospital
data['hospital'] = [1 if gravity == 3 else 0 for gravity in data['gravity']]

# Sum of in hospital per accident
data['hospital_n'] = data.groupby('num_acc')['hospital'].transform('sum')

# applying extra weight
data['severity_score'] = (
    data['severity_score']
    + (data['killed_n'] * 6) ** 0.4
    + (data['hospital_n'] * 3.5) ** 0.2
)

Now, I just have to drop the all features related to gravity except for `severity_score`. Furthermore, I can now just keep one line per accident. 

data.drop(
    columns=[
        'gravity', 'grav_weighted', 'killed', 'killed_n', 'hospital',
        'hospital_n'
    ],
    axis=1, 
    inplace=True
)

data.drop_duplicates(subset='num_acc', inplace=True)

data.head()

# Features

## Latitude and Longitude (main interest variables)

I already perceived problems in the `lat` and `long` features: 1) showing a wrong decimal sign, 2) some missing values are coded with `-`. In the first case, I just need to replace the `,` with `.`, in the second case I need to set a `-` to missing. 

data['long'].unique()

data['long'] = data['long'].astype(str)

data['long'] = data['long'].str.replace(',', '.', regex=False)

data.loc[data['long']=='-', 'long'] = np.nan

data['long'] = data['long'].astype(float)


Now, I do the same steps for latitude.

data['lat'].unique()

data['lat'] = data['lat'].astype(str)

data['lat'] = data['lat'].str.replace(',', '.', regex=False)

data.loc[data['lat']=='-', 'lat'] = np.nan

data['lat'] = data['lat'].astype(float)


The problem is that it seems to be mixed of degrees and some other projected coordinate system (PCS) in the data, since some data falls in the range of lattitude (-90 to +90) and longitude (-180 to +180) in degrees, while some data is clearly out of bound (i.e., 50334). As you can see below, there are 736791 cases where this is the case. 

print(data['lat'][((data['lat'] > 90) | (data['lat'] < -90)) & ((data['long'] > 180) | (data['long'] < -180))].count())

Since there might be different approaches in typing in the lattitude and longitude data, I tried recoding with `geopandas` from different standards that might be used (EPSG:32631, EPSG:2154, ESPG:3395, and EPSG:27572), but none yielded in France. I inspected the transformations visually in `09-visualizations.ipynb` which showed incorrected data points. 
Hence, I filtered the data set to have only data within the limits of lattitude and longitude. This yielded to a correct amount of data within France. 

# # create invalid df with only cases out of bound
# invalid = data[
#     (data['lat'] > 90) | (data['lat'] < -90) |
#     (data['long'] > 180) | (data['long'] < -180)
# ]

# # create valid df with the rest of the data
# valid = data[~data.index.isin(invalid.index)] 

# # assuming invalid rows in meters (EPSG:3395), create a GeoDataFrame
# geom_invalid = [Point(xy) for xy in zip(invalid['long'], invalid['lat'])]
# gdf_invalid = gpd.GeoDataFrame(invalid, geometry=geom_invalid)

# # setting CRS for the invalid data (assuming they are in EPSG:2154)
# gdf_invalid.set_crs(epsg=32631, inplace=True)

# # reprojecting to WGS84 (EPSG:4326)
# gdf_invalid = gdf_invalid.to_crs(epsg=4326)

# # extracting latitude and longitude from GeoDataFrame
# gdf_invalid['lat'] = gdf_invalid.geometry.y
# gdf_invalid['long'] = gdf_invalid.geometry.x

# # combining valid data with reprojected data
# data = pd.concat([valid, gdf_invalid], axis=0)

data.loc[
    (data['lat'] < -90) | (data['lat'] > 90) | 
    (data['long'] < -180) | (data['long'] > 180), 
    ['lat', 'long']
] = np.nan

data.dropna(subset=['long', 'lat'], inplace=True)

data = data[data['lat'] != 0]

Now, I lastly check if any data point is still outside of France.

print(data['lat'][((data['lat'] < 41.3037) | (data['lat'] > 51.1242)) & ((data['long'] < -5.3163) | (data['long'] > 9.6625))].count())

Although we get 0 above, on the map in `09-visualizations.ipynb`, I saw one point in the gulf of Guinea which is clearly not mainland France. This might be a data point, that is plotted by degrees, however the input might be different, and therefore create wrong points which would mess with the algorithms. Since the original format is not given and the trials above did not lead to a sufficient solution, there is only deleting. 

print(data['lat'][((data['lat'] >= 3.5) & (data['lat'] <= 5.5)) & ((data['long'] >= -0.1) & (data['long'] <= 0.1))].count())

data = data[
    ~(((data['lat'] >= 3.5) & (data['lat'] <= 5.5)) & 
    ((data['long'] >= -0.1) & (data['long'] <= 0.1)))
]

Let's check for the remaining features, if any feature has high missing values. Features with missings over 40% will be deleted.

data.isna().mean() * 100

Here it's the case for:

- `road_num_index`
- `v2`
- `max_speed`

All other features will be inspected, probably more features will be deleted, since no information is sometimes coded as `0` or `-1` which is equal to NA but not detected above. Features with low NA's will be handled in feature engineering after test-train split.

## `light_condit`

I turned `-1` no information to `NA` (3 cases).

print(data['light_condit'].value_counts(dropna=False))

data.loc[data['light_condit'] == -1, 'light_condit'] = np.nan

## `location_type`

All good!

print(data['location_type'].value_counts(dropna=False))

## `intersect_type`

Values with `-1`/`0` no information turned to `NA` (17 cases). A lot of categories, further check in feature engineering.

print(data['intersect_type'].value_counts(dropna=False))

# no information to NA
data.loc[data['intersect_type'] == -1, 'intersect_type'] = np.nan
data.loc[data['intersect_type'] == 0, 'intersect_type'] = np.nan

print(data['intersect_type'].value_counts(dropna=False))


## `weather_cond`

Values with `-1` turned to `NA` (3 cases). A lot of categories, further check in feature engineering. 

print(data['weather_cond'].value_counts(dropna=False))

data.loc[data['weather_cond'] == -1, 'weather_cond'] = np.nan

print(data['weather_cond'].value_counts(dropna=False))


## `collision_type`

Cases with no information (`-1`) turned to `NA` (65 cases). A lot of categories, further check in feature engineering.

print(data['collision_type'].value_counts(dropna=False))

data.loc[data['collision_type'] == -1, 'collision_type'] = np.nan

print(data['collision_type'].value_counts(dropna=False))


## `municip_code`

Code for the municipial region based on INSEE. Regard that this correlates with the `dep` variable which has the department code based on INSEE. I explore this further in the feature engineering. There is one case with `N/C` value which is turned to `NA`.

Usually, the INSEE code for municipialities has 5 numbers, and the first two indicate the department and the last 3 the municipalities. However, as you can see below, majority of the entries here is lower than $10000$, indicating an odd format for INSEE for municipalities. Hence, I drop this variable and might only include `dep` next to the lattitude and longitude. 

print(data['municip_code'].value_counts(dropna=False))

data.loc[data['municip_code']=='N/C', 'municip_code'] = np.nan

data['municip_code'] = data['municip_code'].astype('Int64')


insee_test = data['municip_code'] < 10000

insee_test.value_counts()

## `adr`

This information is also captured by longitude and lattitude with more precision. It captures street names (categorical) and has 447383 unique values. I drop this variable, since the information with longitude and lattitude can be better used in the algorithms. 

data['adr'].nunique()

## `gps_code`

This variable has only one level, since I only deal with mainland France. I drop it. 

print(data['gps_code'].value_counts(dropna=False))

print(data['gps_code'].unique())


## `dep`

Inconsistent data. Regarding [government pages](https://www.insee.fr/fr/information/7766585), code for department only have 2 digits, except for oversea territories which have digits over 970. The variable also has geographical information since each department can be represented in an array of longitude and longitude, therefore, I drop the feature here. 

print(data['dep'].value_counts(dropna=False))

print(data['dep'].sort_values().unique())

## `month`

Stays at it is.

print(data['month'].value_counts(dropna=False))

## `day`

stays as it is!

print(data['day'].value_counts(dropna=False))

This variable does not give a lot of information, because a day is very different among months and years. What could be interesting, to have a variable showing which day of the week it was assuming on weekdays accidents might be more frequent/severe.
Therefore, I recode `day` into `weekday`.

data['date'] = pd.to_datetime(data[['year', 'month', 'day']])

data['weekday'] = data['date'].dt.weekday

data['weekday'].value_counts(dropna=False)

## `time`

Some values have still `:`. For consistency, and hour extraction, I delete the `:`. Furthermore, some values only have two digits, sometimes above 24. Probably, these values indicate 0:43. None of these values if above 59, hence, it seems senseful to recode these with a leading 0 for midnight.

print(data['time'].value_counts(dropna=False))

# delete `:` in time
def del_colon(value):
    if ':' in (str(value)):
        new = str(value).replace(':', '')
        return new
    else:
        return value

data['time'] = data['time'].apply(del_colon)

print(data['time'].value_counts(dropna=False))

def get_hours(value):
    if len(str(value)) <= 4 and len(str(value)) >= 3:
        if str(value).startswith('0'):
            return int(str(value)[1:-2])
        else:
            return int(value) // 100
    elif len(str(value)) <= 2 and len(str(value)) >= 1:
        return 0
    elif len(str(value)) > 4:
        print(value)
    else:
        print(value)
    
data['hour'] = data['time'].apply(get_hours)

print(data['hour'].value_counts(dropna=False))

data['time'] = data['time'].astype(float)

print(data['time'][(data['time'] < 100) & (data['time'] > 59)])

## `year`

Stays as it is.

print(data['year'].value_counts(dropna=False))

## `road_cat`

Has a lot of categories, will be checked in feature engineering.

print(data['road_cat'].value_counts(dropna=False))

## `road_num`

In regard to longitude and lattitude it has no new information, therefore I drop it. Usually road numbers have a prefix like A for highways and N for long distance national roads, D for department roads, C for municipal roads and E for European roads. With just the number, it is not clearly separable.

print(data['road_num'].value_counts(dropna=False))

## `traffic_dir`

Keep it so far.

print(data['traffic_dir'].value_counts(dropna=False))

data.loc[data['traffic_dir'] == -1, 'traffic_dir'] = np.nan
data.loc[data['traffic_dir'] == 0, 'traffic_dir'] = np.nan

print(data['traffic_dir'].value_counts(dropna=False))

data['traffic_dir'].isna().mean()

## `num_lanes`

It should indicate the number of road lanes where the accident happened. However, it has quite odd numbers over 12 (Champs Elysee has 10), however, these are few cases. Furthermore, there are some strings, which were recoded to `NA`, too. With the other `NA's` these might be recoded in the feature engineering.

print(data['num_lanes'].value_counts(dropna=False))

def set_int(value):
    try:
        return float(value)
    except ValueError:
        return np.nan

data['num_lanes'] = data['num_lanes'].apply(set_int)

data.loc[data['num_lanes'] == -1, 'num_lanes'] = np.nan
data.loc[data['num_lanes'] == 0, 'num_lanes'] = np.nan
data.loc[data['num_lanes'] == ' -1', 'num_lanes'] = np.nan

print(data['num_lanes'].value_counts(dropna=False))

data['num_lanes'].isna().mean()

## `pr`

After a first easy correction from no information to NA, crossed threshold of more than 30% NA's. Therefore, I dropped.

print(data['pr'].value_counts(dropna=False))

data.loc[data['pr'] == ' -1', 'pr'] = np.nan
data.loc[data['pr'] == '0', 'pr'] = np.nan

data['pr'].isna().mean()

## `pr1`

Dropped as above, since there are more than 50% without information or NA.

print(data['pr1'].value_counts(dropna=False))

data.loc[data['pr1'] == ' -1', 'pr1'] = np.nan
data.loc[data['pr1'] == '0', 'pr1'] = np.nan

data['pr1'].isna().mean()

## `res_lane`

Has a lot of `NA's` (including no information `0`/`-1`). Nearly 93% have NA or no information, therefore I dropped the variable.

print(data['res_lane'].value_counts(dropna=False))

data.loc[data['res_lane'] == 0, 'res_lane'] = np.nan
data.loc[data['res_lane'] == -1, 'res_lane'] = np.nan

print(data['res_lane'].value_counts(dropna=False))

data['res_lane'].isna().mean()

## `prof_road`

Has a lot of `NA's`. Will be further inspected in feature engineering.

print(data['prof_road'].value_counts(dropna=False))

data.loc[data['prof_road'] == 0, 'prof_road'] = np.nan
data.loc[data['prof_road'] == -1, 'prof_road'] = np.nan

print(data['prof_road'].value_counts(dropna=False))

data['prof_road'].isna().mean()

## `plan_view`

Has a lot of `NA's`. Will be further inspected in feature engineering.

print(data['plan_view'].value_counts(dropna=False))

data.loc[data['plan_view'] == 0, 'plan_view'] = np.nan
data.loc[data['plan_view'] == -1, 'plan_view'] = np.nan

print(data['plan_view'].value_counts(dropna=False))

data['plan_view'].isna().mean()

## `tpc_length`

Has a lot of `NA's`/no information. Also super high values over 200m. Due to high missing/no information (~91%) dropped.

print(data['tpc_length'].value_counts(dropna=False))

data.loc[data['tpc_length'] == '0.0', 'tpc_length'] = np.nan
data.loc[data['tpc_length'] == -1, 'tpc_length'] = np.nan

print(data['tpc_length'].value_counts(dropna=False))

data['tpc_length'].isna().mean()

## `width_road`

Some inconsistent values (wrong decimal, too broad roads). The broadest road in France is Avenue Foch with 120 meters. However, lot of NA's and no information values (`0`/`-1`), therefore dropped. Also information is also captured in `num_lanes`, since the number of lanes is directly connected to `width_road`.

print(data['width_road'].value_counts(dropna=False))

data.loc[data['width_road'] == '0.0', 'width_road'] = np.nan
data.loc[data['width_road'] == -1, 'width_road'] = np.nan
data.loc[data['width_road'] == " -1", 'width_road'] = np.nan


print(data['width_road'].value_counts(dropna=False))

data['width_road'].isna().mean()

## `surface`

`NA's` will be handled in feature engineering in addition to meaningful groups in this feature.

print(data['surface'].value_counts(dropna=False))

data.loc[data['surface'] == 0, 'surface'] = np.nan
data.loc[data['surface'] == -1, 'surface'] = np.nan

print(data['surface'].value_counts(dropna=False))


## `infra`

Ca. 92% is NA or no information. Dropped.

print(data['infra'].value_counts(dropna=False))

data.loc[data['infra'] == 0, 'infra'] = np.nan
data.loc[data['infra'] == -1, 'infra'] = np.nan

print(data['infra'].value_counts(dropna=False))

data['infra'].isna().mean()

## `acc_loc`

`NA's` will be handled in feature engineering.

print(data['acc_loc'].value_counts(dropna=False))

data.loc[data['acc_loc'] == 0, 'acc_loc'] = np.nan
data.loc[data['acc_loc'] == -1, 'acc_loc'] = np.nan

print(data['acc_loc'].value_counts(dropna=False))

data['acc_loc'].isna().mean()

## Dropping variables + cases

First, I drop cases, where I don't have information on longitude or lattitude (`NA` or `0`). Next I drop the variables I decided above: `municip_code`, `gps_code`, and `adr`. The cases `0` occured in lattitude and this is not meaningful, since lattitude 0 is the equator. In addition, I delete the identifier `num_acc`. 

data.drop(
    columns=[
        'municip_code', 'adr', 'gps_code', 'dep', 'road_num', 
        'road_num_index', 'v2', 'max_speed', 'pr', 'pr1',
        'res_lane', 'tpc_length', 'width_road', 'infra', 
        'time', 'day', 'num_acc',         
    ], 
    axis=1, 
    inplace=True
)

# Save data frame

data.to_csv('data/processed/clean.csv', index=False)
