==> 03_train_test_split.ipynb <==
# Train-Test split

Here I just do the train-test split before engineering the features in the next script.

# Libraries used
import pandas as pd
from sklearn.model_selection import train_test_split

# loading data
data = pd.read_csv('data/processed/clean.csv', low_memory=False)

X = data.drop(columns='severity_score', axis=1)

y = data['severity_score']

# train-test split
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.25)

train = pd.concat([X_train, y_train], axis=1)
test = pd.concat([X_test, y_test], axis=1)

# saving
train.to_csv('data/processed/train_raw.csv')
test.to_csv('data/processed/test_raw.csv')
